requirements.txt
requests==2.32.3
bs4==0.0.2
python-telegram-bot==22.0
numpy=2.2.5
pandas=2.2.3
dotenv==0.9.9